refactor: 스크롤 상태와 스크롤 제어 로직을 분리 #273
